requirements.txt
httpx[http2]
pynacl
python-dotenv
websockets
